+++ extensions/bitwarden/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Bitwarden Changelog
 
+## [New Command] - 2025-10-07
+
+- Add Create Login command to create new login items
+
 ## [Improvements] - 2025-09-02
 
 - Sync vault on command launch