+++ extensions/summarize-youtube-video-with-ai/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Summarize YouTube Video Changelog
 
+## [Fix] - 2025-10-10
+
+- Update the way transscriptions is handled
+
 ## [Fixes] - 2025-06-27
 
 - 📚 Remove `youtube-transcript` package and install `youtube-transcript-api`